# Improve training time performance