Took a closer look at the ~3min training time:

~ 45s: Upgrade all System Packages
~ 30s: Install R System Dependencies
~ 45s: Upgrade TensorFlow
~ 10s: Initialize Python
~ 10s: Restore R packages from cache
~ 5s: Download train examples, train, upload results
~ 40s: Tearing down TensorFlow.

This PR proposes removing the following stages:

~ 45s: Upgrade all System Packages
~ 45s: Upgrade TensorFlow

Upgrading TensorFlow to use tfestimators uncovered a hang in both Python and R that CloudML needs to resolve. Upgrading all system packages shouldn't be required in most cases since we can rely on CloudML image having a recent version.

<img width="713" alt="screen shot 2017-11-27 at 4 49 23 pm" src="https://user-images.githubusercontent.com/3478847/33296906-fc58bc62-d392-11e7-9af9-dfc194b5400e.png">

to

<img width="714" alt="screen shot 2017-11-27 at 4 49 32 pm" src="https://user-images.githubusercontent.com/3478847/33296912-019a4bd2-d393-11e7-83a2-141c50880b26.png">